+++ src/main/java/com/nisovin/shopkeepers/ui/defaults/HiringHandler.java
@@ -9,8 +9,8 @@ import org.bukkit.inventory.InventoryView;
 import com.nisovin.shopkeepers.Settings;
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
-import com.nisovin.shopkeepers.ui.AbstractUIType;
 import com.nisovin.shopkeepers.ui.AbstractShopkeeperUIHandler;
+import com.nisovin.shopkeepers.ui.AbstractUIType;
 import com.nisovin.shopkeepers.util.Log;
 import com.nisovin.shopkeepers.util.PermissionUtils;
 import com.nisovin.shopkeepers.util.TextUtils;
@@ -22,12 +22,14 @@ public abstract class HiringHandler extends AbstractShopkeeperUIHandler {
 	}
 
 	@Override
-	protected boolean canOpen(Player player) {
+	protected boolean canOpen(Player player, boolean silent) {
 		assert player != null;
 		// Check for hire permission:
 		if (!PermissionUtils.hasPermission(player, ShopkeepersPlugin.HIRE_PERMISSION)) {
-			Log.debug(() -> "Blocked hire window opening for " + player.getName() + ": Missing hire permission.");
-			TextUtils.sendMessage(player, Settings.msgMissingHirePerm);
+			if (!silent) {
+				Log.debug(() -> "Blocked hire window opening for " + player.getName() + ": Missing hire permission.");
+				TextUtils.sendMessage(player, Settings.msgMissingHirePerm);
+			}
 			return false;
 		}
 		return true;